Refactor `user_id` column type from `unsignedInteger` to `unsignedBigInteger`

+++ migrations/2019_02_06_174631_make_acl_rules_table.php
@@ -15,7 +15,7 @@ class MakeAclRulesTable extends Migration
     {
         Schema::create('acl_rules', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->string('disk');
             $table->string('path');
             $table->tinyInteger('access');
